BUG FIX: Independently adjust sibGWAS to have matching effect allele
Previously sibGWAS and population GWAS were assumed to have the same effect and alternate allele. I've fixed this such that this runs an independent check and realignment (esp. applicable when using our sibGWAS with external population GWAS). Also updated the example file to comply with the requirement for allele notation.